Module panorama::config

Module for setting up config files and watchers.

One of the primary goals of panorama is to be able to always hot-reload configuration files.

Structs

Config
Configuration

ImapConfig
Configuring an IMAP server

MailAccountConfig
Configuration for a single mail account

Enums

ImapAuth
Method of authentication for the IMAP server

TlsMethod
Describes when to perform the TLS handshake

Functions

read_config

spawn_config_watcher_system
Start the entire config watcher system, and return a ConfigWatcher, which is a cloneable receiver of config update events.

start_inotify_stream

Type Definitions

ConfigWatcher
Alias for a MailConfig receiver.