Trait panorama::ui::input::HandlesInput

pub trait HandlesInput: Any + Debug + Downcast {
    fn handle_key(
        &mut self, 
        term: &'_ mut Terminal<CrosstermBackend<&'_ mut Stdout>>, 
        evt: KeyEvent
    ) -> Result<InputResult> { ... }
}

Provided methods

fn handle_key(
&mut self,
term: &'_ mut Terminal<CrosstermBackend<&'_ mut Stdout>>,
evt: KeyEvent
) -> Result<InputResult>

Loading content...
Implementations

impl dyn HandlesInput

pub fn is<__T: HandlesInput>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: HandlesInput>(
self: Box<Self>
) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: HandlesInput>(
self: Rc<Self>
) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: HandlesInput>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: HandlesInput>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.