Trait panorama::ui::windows::Window

pub trait Window: HandlesInput {
    fn name(&self) -> String;

    fn draw(
        &self, 
        f: &'_ mut Frame<'_, CrosstermBackend<&'_ mut Stdout>>, 
        area: Rect, 
        ui: &UI
    );

    fn draw_inactive(
        &mut self, 
        f: &'_ mut Frame<'_, CrosstermBackend<&'_ mut Stdout>>, 
        area: Rect, 
        ui: &UI
    ) { ... }
}

Required methods

fn name(&self) -> String

fn draw(
&self,
f: &'_ mut Frame<'_, CrosstermBackend<&'_ mut Stdout>>,
area: Rect,
ui: &UI
)

Loading content...
Provided methods

fn draw_inactive(
&mut self,
f: &'_ mut Frame<'_, CrosstermBackend<&'_ mut Stdout>>,
area: Rect,
ui: &UI
)

Draw function, except the window is not the actively focused one

By default, this just calls the regular draw function but the window may choose to perform a less intensive draw if it's known to not be active

Loading content...
Implementations

impl dyn Window

pub fn is<__T: Window>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Window>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: Window>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: Window>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: Window>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.